Move txnReducer to v3
