+++ v2/components/RewardsItem/package.json
@@ -14,13 +14,13 @@
     "@snx-v2/parseTxnError": "workspace:*",
     "@snx-v2/stakingCalculations": "workspace:*",
     "@snx-v2/txnLink": "workspace:*",
-    "@snx-v2/txnReducer": "workspace:*",
     "@snx-v2/useApr": "workspace:*",
     "@snx-v2/useClaimRewardsMutation": "workspace:*",
     "@snx-v2/useDebtData": "workspace:*",
     "@snx-v2/useFeePoolData": "workspace:*",
     "@snx-v2/useGetLiquidationRewards": "workspace:*",
     "@snx-v2/useRewardsAvailable": "workspace:*",
+    "@snx-v3/txnReducer": "workspace:*",
     "@storybook/react": "^6.5.15",
     "@tanstack/react-query": "^4.3.4",
     "date-fns": "^2.19.0",
