+++ v2/components/SelfLiquidation/package.json
@@ -15,11 +15,11 @@
     "@snx-v2/parseTxnError": "workspace:*",
     "@snx-v2/stakingCalculations": "workspace:*",
     "@snx-v2/txnLink": "workspace:*",
-    "@snx-v2/txnReducer": "workspace:*",
     "@snx-v2/useDebtData": "workspace:*",
     "@snx-v2/useExchangeRatesData": "workspace:*",
     "@snx-v2/useSelfLiquidationData": "workspace:*",
     "@snx-v2/useSelfLiquidationMutation": "workspace:*",
+    "@snx-v3/txnReducer": "workspace:*",
     "@storybook/react": "^6.5.15",
     "@synthetixio/wei": "workspace:*",
     "@tanstack/react-query": "^4.3.4",
