+++ v2/lib/useClaimRewardsMutation/package.json
@@ -4,9 +4,9 @@
   "main": "index.ts",
   "version": "0.0.2",
   "dependencies": {
-    "@snx-v2/txnReducer": "workspace:*",
     "@snx-v2/useGasOptions": "workspace:*",
     "@snx-v2/useSynthetixContracts": "workspace:*",
+    "@snx-v3/txnReducer": "workspace:*",
     "@tanstack/react-query": "^4.3.4",
     "react": "^18.2.0"
   }
