+++ v2/lib/useSelfLiquidationMutation/useSelfLiquidationMutation.ts
@@ -2,7 +2,7 @@ import { useReducer } from 'react';
 import { useMutation } from '@tanstack/react-query';
 import { useSynthetix } from '@snx-v2/useSynthetixContracts';
 import { useGasOptions } from '@snx-v2/useGasOptions';
-import { initialState, reducer } from '@snx-v2/txnReducer';
+import { initialState, reducer } from '@snx-v3/txnReducer';
 import Wei from '@synthetixio/wei';
 
 const createPopulateTransaction = (Synthetix: ReturnType<typeof useSynthetix>['data']) => {
