+++ v3/lib/txnReducer/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@snx-v2/txnReducer",
+  "name": "@snx-v3/txnReducer",
   "private": true,
   "main": "index.ts",
   "version": "0.0.1"
