+++ v3/lib/useApprove/package.json
@@ -4,7 +4,7 @@
   "main": "index.ts",
   "version": "0.0.2",
   "dependencies": {
-    "@snx-v2/txnReducer": "workspace:*",
+    "@snx-v3/txnReducer": "workspace:*",
     "@snx-v3/useAllowance": "workspace:*",
     "@snx-v3/useBlockchain": "workspace:*",
     "@snx-v3/useGasOptions": "workspace:*",
